+++ features-discover/src/main/java/com/jbrunton/mymovies/features/discover/DiscoverViewStateReducer.kt
@@ -3,11 +3,39 @@ package com.jbrunton.mymovies.features.discover
 import com.jbrunton.async.AsyncResult
 import com.jbrunton.async.getOr
 import com.jbrunton.async.map
+import com.jbrunton.mymovies.entities.models.Genre
+import com.jbrunton.mymovies.entities.models.Movie
 import com.jbrunton.mymovies.libs.ui.livedata.SingleLiveEvent
+import com.jbrunton.mymovies.libs.ui.viewmodels.Interactor
 import com.jbrunton.mymovies.usecases.discover.DiscoverState
+import kotlinx.coroutines.flow.Flow
 
-class DiscoverViewStateReducer(val scrollToGenreResults: SingleLiveEvent<Unit>) {
-    fun reduce(previousState: AsyncResult<DiscoverState>, change: DiscoverStateChange): AsyncResult<DiscoverState> {
+sealed class DiscoverIntent {
+    object Load : DiscoverIntent()
+    data class SelectGenre(val genre: Genre) : DiscoverIntent()
+    data class SelectMovie(val movie: Movie) : DiscoverIntent()
+    object ClearSelectedGenre : DiscoverIntent()
+}
+
+sealed class DiscoverStateChange {
+    data class DiscoverResultsAvailable(val discoverResult: AsyncResult<DiscoverState>) : DiscoverStateChange()
+    data class GenreSelected(val selectedGenre: Genre) : DiscoverStateChange()
+    data class GenreResultsAvailable(val genreResults: AsyncResult<List<Movie>>) : DiscoverStateChange()
+    object SelectedGenreCleared : DiscoverStateChange()
+    object Nothing : DiscoverStateChange()
+}
+
+class DiscoverInteractor(
+        val scrollToGenreResults: SingleLiveEvent<Unit>
+) : Interactor<DiscoverIntent, AsyncResult<DiscoverState>, DiscoverStateChange>() {
+
+    override val initialState: AsyncResult<DiscoverState> = AsyncResult.loading(null)
+
+    override suspend fun actionsFor(intent: DiscoverIntent): Flow<DiscoverStateChange> {
+
+    }
+
+    override fun combine(previousState: AsyncResult<DiscoverState>, change: DiscoverStateChange): AsyncResult<DiscoverState> {
         return when (change) {
             is DiscoverStateChange.DiscoverResultsAvailable -> change.discoverResult
             is DiscoverStateChange.GenreSelected -> previousState.map {
